% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_parallel.R
\name{replace.missing}
\alias{replace.missing}
\title{(Internal) Replace missing values with other values,internally used for
parallelization}
\usage{
replace.missing(X, missing = NA, rep)
}
\arguments{
\item{X}{An input vector}

\item{missing}{A charactors representing a missing value}

\item{rep}{A vector of new values to replace missing values}
}
\value{
A vector with replaced values
}
\description{
(Internal) Replace missing values with other values,internally used for
parallelization
}
