% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_labels}
\alias{sample_labels}
\title{Synthetic dataset containing population labels for the dataset simsnp.}
\format{A vector with 753 elements.}
\usage{
data(example_SNP)
}
\description{
A dataset contains a character vector of 753 elements containing labels or
populations of 753 individuals which they belong. Three populations and
outliers were labeled as "pop1", "pop2", "pop3", and "outlier".
}
\seealso{
\code{\link{simsnp}}
}
\keyword{sample_labels}
