% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kris.R
\docType{package}
\name{KRIS-package}
\alias{KRIS}
\alias{KRIS-package}
\title{Keen and Reliable Interface Subroutines for bioinformatic analysis}
\description{
Provides useful functions which are needed for bioinformatic analysis such as
calculating linear principal components from numeric data and
Single-nucleotide polymorphism (SNP) dataset, calculating fixation index
(Fst) using Hudson method, creating scatter plots in 3 views, handling with
PLINK binary file format, detecting rough structures and outliers using
unsupervised clustering, and calculating matrix multiplication in the faster
way for big data.
}
\details{
The R package \pkg{KRIS} requires the package \pkg{rARPACK}.

Here is the list of functions in the R package \pkg{KRIS}:
\itemize{
\item \code{\link{cal.pc.linear}}
\item \code{\link{cal.pc.projection}}
\item \code{\link{fst.each.snp.hudson}}
\item \code{\link{fst.hudson}}
\item \code{\link{plot3views}}
\item \code{\link{read.bed}}
\item \code{\link{rubikclust}}
\item \code{\link{write.bed}}
\item \code{\link{xxt}}
}

Moreover, here is the list of example datasets in the R package \pkg{KRIS}:
\itemize{
\item \code{\link{simsnp}}
\item \code{\link{sample_labels}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kridsadakorn/kris}
  \item Report bugs at \url{https://github.com/kridsadakorn/kris/issues}
}

}
\author{
\strong{Maintainer}: Kridsadakorn Chaichoompu \email{kridsadakorn@biostatgen.org} (\href{https://orcid.org/0000-0002-8123-3409}{ORCID})

Authors:
\itemize{
  \item Kristel Van Steen (\href{https://orcid.org/0000-0001-9868-5033}{ORCID})
  \item Fentaw Abegaz (\href{https://orcid.org/0000-0002-6294-4360}{ORCID})
  \item Sissades Tongsima (\href{https://orcid.org/0000-0002-1491-1839}{ORCID})
  \item Philip James Shaw
  \item Anavaj Sakuntabhai (\href{https://orcid.org/0000-0001-9099-6937}{ORCID})
  \item Luisa Pereira (\href{https://orcid.org/0000-0002-4271-1527}{ORCID})
}

}
\keyword{internal}
