% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kris.R
\docType{package}
\name{kris-package}
\alias{kris}
\alias{kris-package}
\title{Keen and Reliable Interface Subroutines for bioinformatic analysis}
\description{
The R package Kris (Keen and Reliable InterfaceS for bioinformatic
analysis) is the package providing useful funcions which are needed for
bioinformatic analysis.
}
\details{
The R package \pkg{kris} requires the package \pkg{rARPACK}.

Here is the list of functions in the R package \pkg{kris}:
\itemize{
\item \code{\link{cal.pc.linear}}
\item \code{\link{fst.each.snp.hudson}}
\item \code{\link{fst.hudson}}
\item \code{\link{plot3views}}
\item \code{\link{read.bed}}
\item \code{\link{rubikclust}}
\item \code{\link{write.bed}}
\item \code{\link{xxt}}
}

Moreover, here is the list of example datasets in the R package \pkg{kris}:
\itemize{
\item \code{\link{simsnp}}
\item \code{\link{sample_labels}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/kris.ccp/kris_rel}
  \item Report bugs at \url{https://gitlab.com/kris.ccp/kris_rel/issues}
}

}
\author{
\strong{Maintainer}: Kridsadakorn Chaichoompu \email{kridsadakorn.chaichoompu@gmail.com}

Authors:
\itemize{
  \item Kristel Van Steen
  \item Fentaw Abegaz
  \item Sissades Tongsima
  \item Philip James Shaw
  \item Anavaj Sakuntabhai
  \item Luisa Pereira
}

}
\keyword{internal}
